代码增加版权声明。wfVersion和ruleVersion字段的类型修改为long

+++ nop-rule/nop-rule-api/src/main/java/io/nop/rule/api/beans/_gen/_RuleKeyBean.java
@@ -21,19 +21,19 @@
             this._ruleName = value;
         }
     
-        private Integer _ruleVersion;
+        private Long _ruleVersion;
 
         /**
          * 规则版本 
          */
-        public Integer getRuleVersion(){
+        public Long getRuleVersion(){
             return _ruleVersion;
         }
 
         /**
          * 规则版本 
          */
-        public void setRuleVersion(Integer value){
+        public void setRuleVersion(Long value){
             this._ruleVersion = value;
         }
