+++ nop-rule/nop-rule-core/src/main/java/io/nop/rule/core/IRuleManager.java
@@ -1,3 +1,10 @@
+/**
+ * Copyright (c) 2017-2023 Nop Platform. All rights reserved.
+ * Author: canonical_entropy@163.com
+ * Blog:   https://www.zhihu.com/people/canonical-entropy
+ * Gitee:  https://gitee.com/canonical-entropy/nop-chaos
+ * Github: https://github.com/entropy-cloud/nop-chaos
+ */
 package io.nop.rule.core;
 
 import io.nop.commons.cache.ICache;
@@ -12,13 +19,13 @@ public interface IRuleManager {
 
     IRuleRuntime newRuntime();
 
-    IExecutableRule getRule(String ruleName, Integer ruleVersion);
+    IExecutableRule getRule(String ruleName, Long ruleVersion);
 
-    RuleModel getRuleModel(String ruleName, Integer ruleVersion);
+    RuleModel getRuleModel(String ruleName, Long ruleVersion);
 
-    Map<String, Object> executeRule(String ruleName, Integer ruleVersion, IRuleRuntime ruleRt);
+    Map<String, Object> executeRule(String ruleName, Long ruleVersion, IRuleRuntime ruleRt);
 
-    default Object chooseByRule(String ruleName, Integer ruleVersion, IRuleRuntime ruleRt) {
+    default Object chooseByRule(String ruleName, Long ruleVersion, IRuleRuntime ruleRt) {
         Map<String, Object> outputs = executeRule(ruleName, ruleVersion, ruleRt);
         if (outputs == null || !ruleRt.isRuleMatch())
             return null;
